chore(cleanup): remove agent/ — migrate to core/agent and domain/compute
- Delete legacy agent/.cloudbuild/, .gitignore, app/__init__.py,
  tests/conftest.py (old venv path hack)
- Move retained unit tests → core/tests/unit/
- Move arkhai integration tests → domain/compute/tests/integration/
- Move training configs/README → domain/compute/training/
- Add domain/compute/tests/conftest.py (repo-root sys.path)
- Remove "../agent" from core/pyproject.toml pythonpath (was causing
  agent/app/ to shadow core/agent/app/ imports)
- Drop stale test methods: test_validation.py, test_action_executor.py
  (old provision_machine API), TestTokenFromDict (service TOKEN_REGISTRY),
  2 fulfill tests in test_counterparty_url_chain.py (old API)
