--- agent/.cloudbuild/deploy-to-prod.yaml
@@ -1,38 +1,0 @@
-# Copyright 2025 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-steps:
-   
-  - name: "gcr.io/cloud-builders/gcloud"
-    id: trigger-deployment
-    entrypoint: gcloud
-    args:
-      - "run"
-      - "deploy"
-      - "a2a-agent"
-      - "--image"
-      - "$_REGION-docker.pkg.dev/$PROJECT_ID/$_ARTIFACT_REGISTRY_REPO_NAME/$_CONTAINER_NAME"
-      - "--region"
-      - "$_REGION"
-      - "--project"
-      - $_PROD_PROJECT_ID
-
-substitutions:
-  _PROD_PROJECT_ID: YOUR_PROD_PROJECT_ID
-  _REGION: asia-southeast1
-
-logsBucket: gs://${PROJECT_ID}-a2a-agent-logs/build-logs
-options:
-  substitutionOption: ALLOW_LOOSE
-  defaultLogsBucketBehavior: REGIONAL_USER_OWNED_BUCKET
