--- agent/.cloudbuild/staging.yaml
@@ -1,154 +1,0 @@
-# Copyright 2025 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-steps:
-  # Download trained model artifacts from GitHub Release
-  - name: "gcr.io/cloud-builders/git"
-    id: download-models
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        apt-get update && apt-get install -y gh
-        mkdir -p agent/app/policies/models
-        gh release download "${_MODEL_VERSION}" \
-          --repo "${_GITHUB_REPO}" \
-          --pattern "*.pt" \
-          --dir agent/app/policies/models/ \
-          --clobber || echo "WARNING: Model download failed, continuing without models"
-    secretEnv: ['GITHUB_TOKEN']
-
-  # Build and Push
-  - name: "gcr.io/cloud-builders/docker"
-    args:
-      [
-        "build",
-        "-t",
-        "$_REGION-docker.pkg.dev/$PROJECT_ID/$_ARTIFACT_REGISTRY_REPO_NAME/$_CONTAINER_NAME",
-        "--build-arg",
-        "COMMIT_SHA=$COMMIT_SHA",
-        ".",
-      ]
-  - name: "gcr.io/cloud-builders/docker"
-    args:
-      [
-        "push",
-        "$_REGION-docker.pkg.dev/$PROJECT_ID/$_ARTIFACT_REGISTRY_REPO_NAME/$_CONTAINER_NAME",
-      ]
-
-  # Deploy to Staging
-  - name: "gcr.io/cloud-builders/gcloud"
-    id: deploy-staging
-    entrypoint: gcloud
-    args:
-      - "run"
-      - "deploy"
-      - "a2a-agent"
-      - "--image"
-      - "$_REGION-docker.pkg.dev/$PROJECT_ID/$_ARTIFACT_REGISTRY_REPO_NAME/$_CONTAINER_NAME"
-      - "--region"
-      - "${_REGION}"
-      - "--project"
-      - "${_STAGING_PROJECT_ID}"
-
-  # Fetch Staging Service URL
-  - name: "gcr.io/cloud-builders/gcloud"
-    id: fetch-staging-url
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        echo $(gcloud run services describe a2a-agent \
-        --region ${_REGION} --project ${_STAGING_PROJECT_ID} --format="value(status.url)") > staging_url.txt
-
-  # Fetch ID Token
-  - name: gcr.io/cloud-builders/gcloud
-    id: fetch-id-token
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        echo $(gcloud auth print-identity-token -q) > id_token.txt
-
-  # Load Testing
-  - name: "python:3.11-slim"
-    id: load_test
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        export _ID_TOKEN=$(cat id_token.txt)
-        export _STAGING_URL=$(cat staging_url.txt)
-        pip install locust==2.31.1 --user
-        locust -f tests/load_test/load_test.py \
-        --headless \
-        -H $$_STAGING_URL \
-        -t 30s -u 10 -r 0.5 \
-        --csv=tests/load_test/.results/results \
-        --html=tests/load_test/.results/report.html
-    env:
-      - 'PATH=/usr/local/bin:/usr/bin:~/.local/bin'
-
-  # Export Load Test Results to GCS
-  - name: gcr.io/cloud-builders/gcloud
-    id: export-results-to-gcs
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        export _TIMESTAMP=$(date +%Y%m%d-%H%M%S)
-        gsutil -m cp -r tests/load_test/.results gs://${_BUCKET_NAME_LOAD_TEST_RESULTS}/results-$${_TIMESTAMP}
-        echo "_________________________________________________________________________"
-        echo "Load test results copied to gs://${_BUCKET_NAME_LOAD_TEST_RESULTS}/results-$${_TIMESTAMP}"
-        echo "HTTP link: https://console.cloud.google.com/storage/browser/${_BUCKET_NAME_LOAD_TEST_RESULTS}/results-$${_TIMESTAMP}"
-        echo "_________________________________________________________________________"
-
-  # Trigger Prod Deployment
-  - name: gcr.io/cloud-builders/gcloud
-    id: trigger-prod-deployment
-    entrypoint: gcloud
-    args:
-      - "beta"
-      - "builds"
-      - "triggers"
-      - "run"
-      - "deploy-a2a-agent"
-      - "--region"
-      - "$LOCATION"
-      - "--project"
-      - "$PROJECT_ID"
-      - "--sha"
-      - $COMMIT_SHA
-
-  - name: gcr.io/cloud-builders/gcloud
-    id: echo-view-build-trigger-link
-    entrypoint: /bin/bash
-    args:
-      - "-c"
-      - |
-        echo "_________________________________________________________________________"
-        echo "Production deployment triggered. View progress and / or approve on the Cloud Build Console:"
-        echo "https://console.cloud.google.com/cloud-build/builds;region=$LOCATION"
-        echo "_________________________________________________________________________"
-
-substitutions:
-  _STAGING_PROJECT_ID: YOUR_STAGING_PROJECT_ID
-  _REGION: asia-southeast1
-  _MODEL_VERSION: "latest"
-  _GITHUB_REPO: "arkhai-io/simple-market-service"
-
-logsBucket: gs://${PROJECT_ID}-a2a-agent-logs/build-logs
-options:
-  substitutionOption: ALLOW_LOOSE
-  defaultLogsBucketBehavior: REGIONAL_USER_OWNED_BUCKET
