+++ core/pyproject.toml
@@ -109,7 +109,7 @@ build-backend = "hatchling.build"
 
 
 [tool.pytest.ini_options]
-pythonpath = [".", "..", "../agent"]
+pythonpath = [".", ".."]
 asyncio_default_fixture_loop_scope = "function"
 
 [tool.hatch.metadata]
