+++ core/tests/unit/test_price_negotiation.py
