+++ domain/compute/training/config/single_agent_buyer.ini
